requirements.txt
Django==1.10.3
psycopg2==2.6.2
djangorestframework==3.5.3
django-cors-headers==1.3.1
ipython==5.1.0
